att float32

+++ models/resnet.py
@@ -292,10 +292,10 @@ class AttentionBlock(tf.keras.layers.Layer):
         self.dropout = tf.keras.layers.Dropout(0.5)        
         
     def position_encoding(self, d, n):    
-        vals_i = tf.reshape(tf.range(d), (1,-1))
+        vals_i = tf.cast(tf.reshape(tf.range(d), (1,-1)), tf.float32)
         vals_i = tf.tile(vals_i, (n, 1))
-        pos = tf.reshape(tf.range(n), (1,-1))
-        pos = tf.transpose(tf.tile(pos, (d, 1)))        
+        pos = tf.cast(tf.reshape(tf.range(n), (1,-1)), tf.float32)
+        pos = tf.transpose(tf.tile(pos, (d, 1)))                
         sins  = tf.math.sin(pos / tf.math.pow(10000.0, 2.0*vals_i / d))
         cosins  = tf.math.cos(pos / tf.math.pow(10000.0, 2.0*vals_i / d))
         pe =  tf.where(tf.equal(tf.math.floormod(vals_i, 2),0), sins, cosins)
